No more use abstract class and interface for athlete class
