+++ src/main/java/specification/entity/Slime.java
@@ -6,7 +6,7 @@ import specification.property.Mass;
 /**
  * 测试实体：史莱姆
  */
-public class Slime extends AbstractAthlete {
+public class Slime extends Athlete {
     public Slime() {
         this(new Mass(5));
     }
